pyproject.toml
[tool.poetry]
name = "bluestom-admission-test"
version = "0.1.0"
description = "This project was made as part of the Bluestorm company's adimission test"
authors = ["Pedro Mesquita <pedrodardengo@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.37"
pytz = "^2022.1"
passlib = "^1.7.4"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
